Renamed PreOptimizer into LogicalPlanPreOptimizer
