+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/plan/logical/LogicalPlan.java
@@ -65,14 +65,6 @@ public abstract class LogicalPlan extends QueryPlan<LogicalPlan> implements Reso
         return stage.ordinal() >= Stage.OPTIMIZED.ordinal();
     }
 
-    public void setPreOptimized() {
-        stage = Stage.PRE_OPTIMIZED;
-    }
-
-    public boolean preOptimized() {
-        return stage.ordinal() >= Stage.PRE_OPTIMIZED.ordinal();
-    }
-
     public void setOptimized() {
         stage = Stage.OPTIMIZED;
     }
